refactor(core): move bot transport and routing to Core
- вынесен BotController в Core/BotController.swift (роутинг апдейтов Telegram)
- вынесен TelegramService в Core/TelegramService.swift (poll, sendMessage, sendDocument)
- оставлен единый TgReplyKeyboard в Core, удалены дубли из других файлов
- обновлён метод sendDocument: ручная сборка multipart/form-data для стабильности
- KudosController очищен от транспорта и логики меню, зарезервирован под доменную фичу Kudos
